+++ src/main/java/com/global/api/entities/CardIssuerResponse.java
@@ -1,12 +1,16 @@
 package com.global.api.entities;
 
+import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.Accessors;
 
 @Accessors(chain = true)
 @Getter
 @Setter
+@AllArgsConstructor
+@NoArgsConstructor
 public class CardIssuerResponse {
     // The result code of the AVS check from the card issuer
     public String avsResult;
@@ -18,4 +22,8 @@ public class CardIssuerResponse {
     public String avsAddressResult;
     // The result of the AVS postal code check from the card issuer
     public String avsPostalCodeResult;
+
+    public CardIssuerResponse(String result){
+        this.result = result;
+    }
 }