+++ src/test/java/com/global/api/tests/portico/PorticoAchTests.java
@@ -5,6 +5,8 @@ import com.global.api.entities.Address;
 import com.global.api.entities.Transaction;
 import com.global.api.entities.enums.*;
 import com.global.api.entities.exceptions.ApiException;
+import com.global.api.entities.exceptions.GatewayException;
+import com.global.api.entities.exceptions.UnsupportedTransactionException;
 import com.global.api.paymentMethods.eCheck;
 import com.global.api.serviceConfigs.PorticoConfig;
 import org.junit.Test;
@@ -18,13 +20,19 @@ import static org.junit.Assert.assertEquals;
 public class PorticoAchTests {
     private eCheck check;
     private Address address;
+    private final String clientTxnID;
+    private static final String TRANSACTION_EXCEPTION = "Either ClientTxnId or GatewayTxnId must be provided for this payment type.";
     
     public PorticoAchTests() throws ApiException {
         PorticoConfig config = new PorticoConfig();
         config.setSecretApiKey("skapi_cert_MTyMAQBiHVEAewvIzXVFcmUd2UcyBge_eCpaASUp0A");
         config.setServiceUrl("https://cert.api2.heartlandportico.com");
+        config.setEnableLogging(true);
 
         ServicesContainer.configureService(config);
+
+        int randomID = new Random().nextInt(999999 - 10000)+10000;
+        clientTxnID = Integer.toString(randomID);
 
         check = new eCheck();
         check.setAccountNumber("1357902468");
@@ -53,6 +61,7 @@ public class PorticoAchTests {
         Transaction response = check.charge(new BigDecimal(11))
                 .withCurrency("USD")
                 .withAddress(address)
+                .withClientTransactionId(clientTxnID)
                 .execute();
         assertNotNull(response);
         assertEquals("00", response.getResponseCode());
@@ -74,7 +83,64 @@ public class PorticoAchTests {
         assertNotNull(voidResponse);
         assertEquals("00", voidResponse.getResponseCode());
     }
+    @Test
+    public void checkQuery_withClientTxnId() throws ApiException {
+        Transaction response = check.charge(new BigDecimal(10))
+                .withCurrency("USD")
+                .withAddress(address)
+                .withClientTransactionId(clientTxnID)
+                .withAllowDuplicates(true)
+                .execute();
+        assertNotNull(response);
+        assertEquals("00", response.getResponseCode());
 
+        Transaction response1 = check.checkQuery()
+                .withCurrency("USD")
+                .withClientTxnId(response.getClientTransactionId())
+                .execute();
+        assertNotNull(response1);
+        assertEquals("00", response1.getResponseCode());
+    }
+    @Test
+    public void checkQuery_WithGatewayTxnId() throws ApiException {
+        Transaction response = check.charge(new BigDecimal(10))
+                .withCurrency("USD")
+                .withAddress(address)
+                .withAllowDuplicates(true)
+                .execute();
+        assertNotNull(response);
+        assertEquals("00", response.getResponseCode());
+
+        Transaction response1 = check.checkQuery()
+                .withGatewayTxnId(response.getTransactionId())
+                .withCurrency("USD")
+                .execute();
+        assertNotNull(response1);
+        assertEquals("00", response1.getResponseCode());
+    }
+    // Negative Test case
+    // without Id
+    // With Incorrect GatewayTxnId
+    @Test
+    public void checkQuery_WithoutId() {
+        UnsupportedTransactionException exc = assertThrows(UnsupportedTransactionException.class,()-> {
+               check.checkQuery()
+                    .withCurrency("USD")
+                    .execute();
+        });
+        assertEquals(TRANSACTION_EXCEPTION,exc.getMessage());
+    }
+    @Test
+    public void checkQuery_randomId() {
+        GatewayException exc = assertThrows(GatewayException.class,()-> {
+            Transaction response = check.checkQuery()
+                    .withGatewayTxnId("5342315375")
+                    .withCurrency("USD")
+                    .execute();
+            assertNotNull(response);
+        });
+        assertEquals("3", exc.getResponseCode());
+    }
     @Test
     public void checkNewCryptoURL() throws ApiException {
         PorticoConfig config = new PorticoConfig();